Refactored to pull report generation out of the main class.
This is in preparation for introducing additional reporting.
