+++ src/test/java/org/onlab/jdvue/ProgArgsTest.java
@@ -1,8 +1,25 @@
+/*
+ * Copyright 2015-present Open Networking Laboratory
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.onlab.jdvue;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 /**
  * Unit tests for program arguments object.
@@ -16,7 +33,6 @@ public class ProgArgsTest {
     @Test
     public void basics() {
         ProgArgs pa = new ProgArgs(new String[]{CAT_PATH});
-//        System.out.println(pa);
 
         assertEquals("wrong catalog path", CAT_PATH, pa.catPath());
         assertFalse("DA flag should be false", pa.dumpAnalysis());
@@ -30,7 +46,6 @@ public class ProgArgsTest {
     @Test
     public void dumpFlag() {
         ProgArgs pa = new ProgArgs(new String[]{CAT_PATH, DUMP_FLAG});
-//        System.out.println(pa);
 
         assertEquals("wrong catalog path", CAT_PATH, pa.catPath());
         assertTrue("DA flag should be true", pa.dumpAnalysis());
@@ -39,7 +54,6 @@ public class ProgArgsTest {
     @Test
     public void otherFlag() {
         ProgArgs pa = new ProgArgs(new String[]{CAT_PATH, OTHER_FLAG});
-//        System.out.println(pa);
 
         assertEquals("wrong catalog path", CAT_PATH, pa.catPath());
         assertFalse("DA flag should be false", pa.dumpAnalysis());
